+++ tokio-util/src/loom.rs
@@ -1,0 +1,1 @@
+pub(crate) use std::sync;
