+++ tokio/src/sync/tests/mod.rs
@@ -7,8 +7,6 @@ cfg_not_loom! {
 cfg_loom! {
     mod loom_atomic_waker;
     mod loom_broadcast;
-    #[cfg(tokio_unstable)]
-    mod loom_cancellation_token;
     mod loom_list;
     mod loom_mpsc;
     mod loom_notify;
